pyproject.toml
[project]
name = "kg-rag"
version = "0.1.0"
description = "Scalable Knowledge-Graph RAG with GraphRAG + LightRAG + HippoRAG2 + Agentic LangGraph"
requires-python = ">=3.11"
authors = [{ name = "Fu Yujiang (Russell)" }]

dependencies = [
  # ---- Web ----
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "sse-starlette>=2.1",
  "httpx>=0.27",
  # ---- LLM / Agents ----
  "anthropic>=0.40",
  "openai>=1.50",
  "litellm>=1.50",
  "langchain>=0.3",
  "langchain-community>=0.3",
  "langgraph>=0.2.40",
  "langgraph-checkpoint-redis>=0.0.5",
  # ---- Retrieval ----
  "qdrant-client>=1.11",
  "neo4j>=5.22",
  "graphdatascience>=1.11",
  "opensearch-py>=2.6",
  "rank-bm25>=0.2",
  "FlagEmbedding>=1.3",
  "sentence-transformers>=3.0",
  # ---- Ingestion ----
  "unstructured[all-docs]>=0.15",
  "docling>=2.0",
  "tiktoken>=0.8",
  "presidio-analyzer>=2.2",
  "presidio-anonymizer>=2.2",
  # ---- Distributed ----
  "ray[default]>=2.34",
  "aiokafka>=0.11",
  "redis>=5.0",
  # ---- MCP ----
  "mcp>=0.9",
  # ---- Observability ----
  "opentelemetry-sdk>=1.26",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "arize-phoenix>=4.30",
  "langfuse>=2.50",
  # ---- Eval ----
  "ragas>=0.2",
  "datasets>=3.0",
  # ---- Utils ----
  "loguru>=0.7",
  "orjson>=3.10",
  "typer>=0.12",
  "tenacity>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.23", "ruff>=0.6", "mypy>=1.11", "locust>=2.31"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
